backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.4
django-admin-sortable==2.3
django-admin-sortable2
django-admin-tailwind==0.0.31
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-js-asset==2.1.0
djangorestframework==3.14.0
docopt==0.6.2
gunicorn==21.2.0
idna==3.4
packaging==23.1
Pillow==10.0.0
pip-check-reqs
pipdeptree==2.13.0
pipreqs==0.4.13
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.6
yarg==0.1.9
